chore: data type of tsnprintf

+++ include/os/osString.h
@@ -81,7 +81,7 @@ typedef struct {
     (dst)[(size)-1] = 0;                 \
   } while (0)
 
-int64_t tsnprintf(char *dst, int64_t size, const char *format, ...);
+int64_t tsnprintf(char *dst, size_t size, const char *format, ...);
 #define TAOS_STRCPY(_dst, _src)         ((void)strcpy(_dst, _src))
 #define TAOS_STRNCPY(_dst, _src, _size) ((void)strncpy(_dst, _src, _size))
 #define TAOS_STRCAT(_dst, _src)         ((void)strcat(_dst, _src))
